Moves enums into a separate folder
