+++ src/components/form.tsx
@@ -5,8 +5,8 @@ import ColourPicker from "./colour-picker"
 import ImageSelect from "./image-select"
 import { useState } from 'react'
 import { Furniture } from '../types/furniture';
-import { HeadlineSize, StandfirstSize } from '../types/size';
-import { Device } from '../types/device';
+import { HeadlineSize, StandfirstSize } from '../enums/size';
+import { Device } from '../enums/device';
 
 export default (props: {furniture?: Furniture, updateFurniture: (newFurniture: Furniture) => void, updateOriginalImageData: (imageData: object) => void }) => {
   const swatchSelectOptions = Object.keys(Config.swatches)
