+++ src/enums/device.ts
